project/aidlc-docs/construction/unit5_card_sharing/src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
pydantic-settings==2.1.0
